Reorganize folder layout
I have kept only one version of the script (the OpenCV version, with the
non-OpenCV code inserted but commented) and renamed it "nanocolor.py".
I have also created "doc" for documentation, and "data" for the data
files.
